+++ R/aio.R
@@ -83,10 +83,12 @@ send_aio <- function(con, data, mode = c("serial", "raw"), timeout = NULL)
 #'     distiguishable from an integer message value). This can be verified using
 #'     \code{\link{is_error_value}}.
 #'
-#'     If the raw message was successfully received but an error occurred in
-#'     unserialisation or data conversion (for example if the incorrect mode was
-#'     specified), the received raw vector will be stored at \code{$data} to
-#'     allow for the data to be recovered.
+#'     For \code{mode = "serial"}, attempting to unserialise a non-serialised
+#'     message will result in the error 'unknown input format'.
+#'
+#'     For all other modes, if an error occurred in conversion of the data to
+#'     the specified mode, a raw vector will be returned at \code{$data} instead
+#'     to allow for the data to be recovered.
 #'
 #' @examples
 #' s1 <- socket("pair", listen = "inproc://nanonext")